+++ lib/reconstructed_app/presentation/widgets/button/button.dart
@@ -1,0 +1,1 @@
+export 'translate_button.dart';
